PDDA-474 Moving getCourtSites() out of JudgeService into AbstractService
